+++ Data Structures/linked.cpp
