+++ General/Diophantine_Equation.cpp
