+++ General/palindrome.java
