+++ Solutions to Known Problems/between two sets.cpp
