+++ Solutions to Known Problems/cats and mouse.cpp
