+++ Solutions to Known Problems/electronics shop.cpp
