+++ Solutions to Known Problems/sock merchant.cpp
